old_requirements.txt
absl-py==1.1.0
addict==2.4.0
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
boto3==1.28.78
botocore==1.31.78
cachetools==5.2.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
# charset-normalizer==2.1.0
click==8.1.7
cmake==3.27.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dlib
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-MySQLdb==2.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
huggingface-hub==0.18.0
humanfriendly==10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jstyleson==0.0.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.3.7
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mediapipe==0.10.7
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
mysqlclient==2.2.0
networkx==3.1
numpy==1.26.1
oauthlib==3.2.0
onnx==1.14.1
onnxruntime==1.16.2
opencv-contrib-python==4.8.1.78
opencv-python==4.5.5.64
opencv-python-headless==4.8.1.78
openvino==2023.1.0
openvino-dev==2023.1.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
proglog==0.1.9
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycparser==2.21
Pygments==2.12.0
pyparsing==2.4.7
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.1
PyYAML==6.0
# requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
# roboflow==1.1.9
rsa==4.9
s3transfer==0.7.0
# sahi==0.11.15
scipy==1.10.1
seaborn==0.11.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sqlparse==0.4.4
stack-data==0.3.0
stow==1.2.2
supervision==0.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
texttable==1.7.0
tf2onnx==1.15.1
thop==0.1.1.post2209072238
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2023.3
# ultralytics==8.0.207
urllib3==1.26.9
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
Werkzeug==3.0.1
wincertstore==0.2
wrapt==1.14.1
yolov5==7.0.13
